requirements.txt
Flask==3.1.0
flask_paginate==2024.4.12
mysql-connector-python==9.1.0
